# Reduce memory usage in read streams