This change avoids allocating a 16KB buffer for every Shadowsocks
Reader.  Instead, the open streams share a pool of buffers,
allocating only as many as are needed, i.e. the number of streams that
currently have data queued for reading.  Most streams are expected
to be idle most of the time, so this should reduce peak memory
requirements on heavily loaded servers and memory-constrained
clients.

The benchmarks confirm that this change does not increase the number of
allocations, and can sometimes lower it:

Master:
```
BenchmarkTCPThroughput-4    78934     16120 ns/op  496 mbps   0 B/op    0 allocs/op
BenchmarkTCPMultiplexing-4   4668    247983 ns/op        121949 B/op  682 allocs/op
```

This branch:
```
BenchmarkTCPThroughput-4    68811     16135 ns/op  496 mbps   0 B/op    0 allocs/op
BenchmarkTCPMultiplexing-    5200    256331 ns/op         86646 B/op  668 allocs/op
```

With an instrumented build on Android, the number of open reads easily
climbed over 200 while browsing, but the number of outstanding buffers
was always below 20, for a savings of ~3 MB.