@bemasc 2020-10-09
An instrumented branch with logging to show the number of open sockets and the number of allocated read buffers is available here: https://github.com/Jigsaw-Code/outline-ss-server/commit/b4b7165862c67ccb00ffa97cdf6eb9d72bd4dcd0