@bemasc 2020-10-09
We need two buffers because idle streams sit blocked in the first Read call
for minutes at a time, waiting for the next chunk.  During that blocked
call, we don't want to be holding the full size buffer.

On Fri, Oct 9, 2020, 1:21 PM Vinicius Fortuna <notifications@github.com>
wrote:

> *@fortuna* requested changes on this pull request.
> ------------------------------
>
> In shadowsocks/cipher.go
> <https://github.com/Jigsaw-Code/outline-ss-server/pull/90#discussion_r502560358>
> :
>
> > @@ -77,6 +77,16 @@ func newAesGCM(key []byte) (cipher.AEAD, error) {
>  	return cipher.NewGCM(blk)
>  }
>
> +func maxCipherOverhead() int {
>
> I prefer the term "TagSize" for consistency with the Shadowsocks
> documentation and the rest of this code.
> ------------------------------
>
> In shadowsocks/stream.go
> <https://github.com/Jigsaw-Code/outline-ss-server/pull/90#discussion_r502562229>
> :
>
> >  )
>
>  // payloadSizeMask is the maximum size of payload in bytes.
>  const payloadSizeMask = 0x3FFF // 16*1024 - 1
>
> +// Buffer pool used for encrypting and decrypting Shadowsocks streams.
> +// The largest buffer we could need is for encrypting a max-length payload.
> +var ssPool = slicepool.MakePool(payloadSizeMask + maxCipherOverhead())
>
> We could have a pool per cipher, but since they all have the same tag
> size, the simpler code doesn't waste any memory.
> ------------------------------
>
> In shadowsocks/stream.go
> <https://github.com/Jigsaw-Code/outline-ss-server/pull/90#discussion_r502562720>
> :
>
> > @@ -262,7 +268,10 @@ type chunkReader struct {
>  	aead cipher.AEAD
>  	// Index of the next encrypted chunk to read.
>  	counter []byte
> -	buf     []byte
> +	// Buffer for the uint16 size and its AEAD tag.  Made in init().
> +	size []byte
>
> Rename to payloadSizeBuf to be clearer
> ------------------------------
>
> In shadowsocks/stream.go
> <https://github.com/Jigsaw-Code/outline-ss-server/pull/90#discussion_r502566746>
> :
>
> > @@ -262,7 +268,10 @@ type chunkReader struct {
>  	aead cipher.AEAD
>  	// Index of the next encrypted chunk to read.
>  	counter []byte
> -	buf     []byte
> +	// Buffer for the uint16 size and its AEAD tag.  Made in init().
> +	size []byte
>
> Why do we need two buffers? You can call Acquire before reading the size.
> Wouldn't it be simpler if we just had a single buffer as before? I find it
> a bit easier to reason about because it's the buffer for the output and it
> all goes together in one write.
> ------------------------------
>
> In shadowsocks/stream.go
> <https://github.com/Jigsaw-Code/outline-ss-server/pull/90#discussion_r502570959>
> :
>
> >  	// In Shadowsocks-AEAD, each chunk consists of two
>  	// encrypted messages.  The first message contains the payload length,
> -	// and the second message is the payload.
> -	sizeBuf := cr.buf[:2+cr.aead.Overhead()]
> -	if err := cr.readMessage(sizeBuf); err != nil {
> +	// and the second message is the payload.  Idle read threads will
> +	// block here until the next chunk.
> +	if err := cr.readMessage(cr.size); err != nil {
>
> Perhaps call Acquire here, so you don't need two buffers?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/Jigsaw-Code/outline-ss-server/pull/90#pullrequestreview-505846454>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AABO3SNWHABLZNN2HYP76ADSJ5BCVANCNFSM4SKH46EQ>
> .
>